[Feat] Main -> Extra로 코드 이동
