+++ janghakhere/janghakhere/Feature/Onboarding/OnboardingMain/OnboardingMainView.swift
@@ -531,251 +531,6 @@ extension OnboardingMainView {
         }
         .paddingHorizontal()
     }
-    
-    @ViewBuilder
-    func militaryServiceContent() -> some View {
-        VStack(alignment: .leading, spacing: 0) {
-            HStack {
-                Spacer()
-                Button {
-                    // TODO: 다음에 입력
-                } label: {
-                    Text("다음에 입력할래요")
-                        .foregroundStyle(.gray400)
-                        .font(.text_sm)
-                        .underline()
-                }
-            }
-            .padding(.top, 16)
-            
-            Text("군대를 다녀오셨나요?")
-                .font(.title_md)
-                .foregroundStyle(.mainGray)
-                .padding(.top, 23)
-                .padding(.bottom, 60)
-            
-            VerticalButtonGroup<MilitaryStatus>(
-                buttonList: MilitaryStatus.allCases, selectedElement: $viewModel.militaryStatus
-            )
-            .padding(.vertical, 8)
-            Spacer()
-            MainButtonView(
-                title: "다음",
-                action: {},
-                disabled: viewModel.militaryStatus == .notSelected
-            )
-        }
-        .paddingHorizontal()
-    }
-    
-    @ViewBuilder
-    func siblingContent() -> some View {
-        VStack(alignment: .leading, spacing: 0) {
-            HStack {
-                Spacer()
-                Button {
-                    // TODO: 다음에 입력
-                } label: {
-                    Text("다음에 입력할래요")
-                        .foregroundStyle(.gray400)
-                        .font(.text_sm)
-                        .underline()
-                }
-            }
-            .padding(.top, 16)
-            Text("형제가 있으신가요?")
-                .font(.title_md)
-                .foregroundStyle(.mainGray)
-                .padding(.top, 23)
-                .padding(.bottom, 60)
-            
-            VerticalButtonGroup<SiblingStatus>(
-                buttonList: SiblingStatus.allCases, selectedElement: $viewModel.siblingStatus
-            )
-            .padding(.vertical, 8)
-            Spacer()
-            MainButtonView(
-                title: "다음",
-                action: {},
-                disabled: viewModel.siblingStatus == .notSelected
-            )
-        }
-        .paddingHorizontal()
-    }
-    
-    @ViewBuilder
-    func incomeDecileContent() -> some View {
-        VStack(alignment: .leading, spacing: 0) {
-            HStack {
-                Spacer()
-                Button {
-                    
-                } label: {
-                    Text("다음에 입력할래요")
-                        .foregroundStyle(.gray400)
-                        .font(.text_sm)
-                        .underline()
-                }
-            }
-            .padding(0)
-            
-            Text("소득구간을 선택해 주세요")
-                .font(.title_md)
-                .foregroundStyle(.mainGray)
-                .padding(.top, 23)
-                .padding(.bottom, 60)
-            
-            GrayBoxGridView<IncomeDecile>(
-                column: .two,
-                titleList: IncomeDecile.allCases,
-                action: {},
-                selectedElement: $viewModel.incomeDecile
-            )
-            .padding(.vertical, 8)
-            
-            Spacer()
-            HStack {
-                Spacer()
-                Button {
-                    viewModel.isShowIncomeSheet = true
-                } label: {
-                    Text("소득구간을 잘 모르겠나요?")
-                        .font(.text_md)
-                        .foregroundStyle(.gray500)
-                        .underline()
-                }
-                .sheet(isPresented: $viewModel.isShowIncomeSheet) {
-                    IncomeConfirmView()
-                }
-                Spacer()
-            }
-            Spacer()
-            
-            MainButtonView(
-                title: "다음",
-                action: {},
-                disabled: viewModel.incomeDecile == .notSelected
-            )
-        }
-        .paddingHorizontal()
-    }
-    
-    
-    @ViewBuilder
-    func academicInfoContent() -> some View {
-        ScrollView(.vertical) {
-            VStack(alignment: .leading, spacing: 0) {
-                HStack {
-                    Spacer()
-                    Button {
-                        // MARK: 다음에 입력할래요
-                    } label: {
-                        Text("다음에 입력할래요")
-                            .foregroundStyle(.gray400)
-                            .font(.text_sm)
-                            .underline()
-                    }
-                }
-                .padding(0)
-                
-                Text("학교 정보를 입력해 주세요")
-                    .font(.title_md)
-                    .foregroundStyle(.mainGray)
-                    .padding(.top, 23)
-                    .padding(.bottom, 31)
-                
-                Text("직전학기 성적")
-                    .font(.title_xsm)
-                    .foregroundStyle(.gray600)
-                    .padding(.vertical, 6)
-                
-                HStack {
-                    GrayLineNumberFieldView(number: $viewModel.previousGrade, maxGradeStatus: $viewModel.maximumGrade)
-                    Text("/")
-                        .font(.title_sm)
-                        .foregroundStyle(.gray300)
-                    HStack(spacing: 12) {
-                        Text(viewModel.maximumGrade == .notApplicable ? "--" : viewModel.maximumGrade.rawValue)
-                            .font(.title_sm)
-                            .foregroundStyle(.black)
-                        Image(systemName: "chevron.down")
-                            .foregroundStyle(.gray500)
-                    }
-                    .padding(.horizontal, 12)
-                    .padding(.vertical, 12)
-                    .background(
-                        RoundedRectangle(cornerRadius: 10)
-                            .fill(.gray60)
-                    )
-                    .onTapGesture {
-                        viewModel.isShowGradeSheet = true
-                    }
-                    .sheet(isPresented: $viewModel.isShowGradeSheet) {
-                        MaxGradeSheet(maxGrade: $viewModel.maximumGrade)
-                    }
-                    Text("점")
-                        .font(.title_xmd)
-                        .foregroundStyle(.black)
-                }
-                .padding(.bottom, 46)
-                
-                Text("전체학기 성적")
-                    .font(.title_xsm)
-                    .foregroundStyle(.gray600)
-                    .padding(.vertical, 7)
-                HStack {
-                    GrayLineNumberFieldView(number: $viewModel.entireGrade, maxGradeStatus: $viewModel.maximumGrade)
-                    Text("/")
-                        .font(.title_sm)
-                        .foregroundStyle(.gray300)
-                    HStack(spacing: 12) {
-                        Text(viewModel.maximumGrade == .notApplicable ? "--" : viewModel.maximumGrade.rawValue)
-                            .font(.title_sm)
-                            .foregroundStyle(.black)
-                        Image(systemName: "chevron.down")
-                            .foregroundStyle(.gray500)
-                    }
-                    .padding(.horizontal, 12)
-                    .padding(.vertical, 12)
-                    .background(
-                        RoundedRectangle(cornerRadius: 10)
-                            .fill(.gray60)
-                    )
-                    .onTapGesture {
-                        viewModel.isShowGradeSheet = true
-                    }
-                    .sheet(isPresented: $viewModel.isShowGradeSheet) {
-                        MaxGradeSheet(maxGrade: $viewModel.maximumGrade)
-                    }
-                    Text("점")
-                        .font(.title_xmd)
-                        .foregroundStyle(.black)
-                }
-                .padding(.bottom, 46)
-                VStack(alignment: .leading, spacing: 0) {
-                    Text("전공계열")
-                        .font(.title_xsm)
-                        .foregroundStyle(.gray600)
-                        .padding(.vertical, 12)
-                    
-                    GrayBoxGridView<MajorField>(
-                        column: .three,
-                        titleList: MajorField.allCases,
-                        action: {},
-                        selectedElement: $viewModel.majorField
-                    )
-                    .padding(.vertical, 8)
-                }
-                Spacer()
-                MainButtonView(
-                    title: "다음",
-                    action: {},
-                    disabled: viewModel.name == ""
-                )
-            }
-        }
-        .paddingHorizontal()
-    }
 }
 
 #Preview {
